requirements.txt
streamlit>=1.31.0
pillow>=10.0.0
groq>=0.9.0
python-dotenv>=1.0.0
